fix: added a star button for the favorites
