+++ client/app/glossary/components/TermListItem.tsx
@@ -2,12 +2,13 @@ import React from "react";
 import { View, Text, TouchableOpacity, StyleProp, ViewStyle } from "react-native";
 import tw from "tailwind-react-native-classnames";
 import Colors from "@/constants/Colors";
-import { ChevronRight } from "lucide-react-native";
+import { ChevronRight, Star } from "lucide-react-native";
 
 export interface TermItem {
   id: string;
   title: string;
   summary: string;
+  isFavorite?: boolean;
 }
 
 interface TermListItemProps {
@@ -17,6 +18,7 @@ interface TermListItemProps {
 }
 
 export default function TermListItem({ item, onPress, containerStyle }: TermListItemProps) {
+  const isFavorite = !!item.isFavorite;
   return (
     <TouchableOpacity
       accessibilityRole="button"
@@ -24,7 +26,15 @@ export default function TermListItem({ item, onPress, containerStyle }: TermList
       onPress={() => onPress && onPress(item)}
     >
       <View style={[tw`rounded-xl px-5 py-5 bg-white flex-row items-start`, { borderColor: "#e5e7eb", borderWidth: 1 }]}> 
-        <View style={[tw`w-2 h-2 rounded-full mt-1 mr-4`, { backgroundColor: "#fbbf24" }]} />
+        <View style={tw`mt-1 mr-4`}>
+          <Star
+            size={18}
+            color={isFavorite ? "#f59e0b" : "#9ca3af"}
+            strokeWidth={2}
+            // Fill when favorited; outline when not
+            fill={isFavorite ? "#f59e0b" : "none"}
+          />
+        </View>
         <View style={tw`flex-1`}>
           <Text style={[tw`font-bold mb-2 text-base`, { color: Colors.text.head }]}>{item.title}</Text>
           <Text style={[tw`text-sm leading-5`, { color: Colors.text.sub }]}>{item.summary}</Text>
